added Gabe's crazy excel spreadsheet thing
We need to test the accuracy of this extensively before using it.  Once
we figure out how to run it, we should make a number of predictions for
various days and see how close we are.  Obviously our ability to test
will be limited since we can't launch anything, but we should test what
we can